Refactor DocumentTypes to web.types to fix circular dependency.
The enum is referenced throughout the system so move it out of
web/utils/pdfs/types.py.
