+++ pii-ner-exclude.txt
@@ -2433,3 +2433,10 @@ Export Applications Without Document Packs
 XXXX
 {{ process.consignment_country
 {{ licence_end_date.strftime("%d %B %
+"Creates a SIL
+f'<span
+PdfGenerator
+"Returns the correct
+DocumentTypes
+AppCls
+doc_type
