+++ web/types.py
@@ -1,3 +1,5 @@
+from enum import Enum, auto
+
 from django.http import HttpRequest
 
 from web.domains.user.models import User
@@ -7,3 +9,15 @@ from web.middleware.common import ICMSMiddlewareContext
 class AuthenticatedHttpRequest(HttpRequest):
     user: User
     icms: ICMSMiddlewareContext
+
+
+class DocumentTypes(Enum):
+    """Types of documents generated by the system"""
+
+    LICENCE_PREVIEW = auto()
+    LICENCE_PRE_SIGN = auto()
+    LICENCE_SIGNED = auto()
+    CERTIFICATE = auto()
+    COVER_LETTER_PREVIEW = auto()
+    COVER_LETTER_PRE_SIGN = auto()
+    COVER_LETTER_SIGNED = auto()
