+++ web/utils/pdf/__init__.py
@@ -1,4 +1,3 @@
 from .generator import PdfGenerator
-from .types import DocumentTypes
 
-__all__ = ["DocumentTypes", "PdfGenerator"]
+__all__ = ["PdfGenerator"]
