+++ web/utils/pdf/generator.py
@@ -9,9 +9,9 @@ from django.template.loader import render_to_string
 from web.domains.case._import.models import ImportApplicationLicence
 from web.domains.case.types import ImpOrExp
 from web.flow.models import ProcessTypes
+from web.types import DocumentTypes
 
 from . import utils
-from .types import DocumentTypes
 
 
 @dataclass
